requirements.txt
pywebview>=5.0
pystray>=0.19
Pillow>=10.0
requests>=2.0
python-dotenv>=1.0
platformdirs>=4.0
